LastMovedFix
Added functionality to loadbalancenodes.sh to verify that the last requested VM migration has completed successfully before attempting to move another VM. Will exit the script if the last requested move has not completed. Will output to shell and rewrite log.txt with the last attempted move if it was skipped.