High/Low Delta
Gets a delta between the CPU usage of the Highest node and the Lowest node. If the difference is greater than or equal to 5%, then migration will commence.

(Ex. node1 is at 10% and node2 is at 5%, migration will complete)
(Ex. node1 is at 10% and node2 is at 12%, migration will skip)